[SET-989] refactor project insights layering
